Delete xml file

+++ model/xml.php
@@ -42,12 +42,14 @@
 		public static function deleteXMLFile($id){
 			$connexion = DBConnexion::connectToDB();
 
-			$query = $connexion->prepare('SELECT * FROM xml WHERE id = :id';
+			$query = $connexion->prepare('SELECT filename FROM xml WHERE id = :id');
 			$query->execute(array('id' => $id));
-
-
-			$query = $connexion->prepare('DELETE FROM xml WHERE id =  :id');
-			$query->execute(array('id' => $id));
+			$result = $query->fetch(PDO::FETCH_ASSOC);
+			$bool = unlink("../xml/".$result['filename']);
+			if($bool === TRUE){
+				$query = $connexion->prepare('DELETE FROM xml WHERE id =  :id');
+				$query->execute(array('id' => $id));
+			}
 
 		 	return $result;
 		}
